🐛 : "auto 슬라이드 자동 재생 기능 수정" #4
