+++ src/components/LandingSlide/AutoSlide.module.css
@@ -1,0 +1,22 @@
+.slideWrapper {
+    display: flex;
+    gap: 20px;
+    padding-bottom: 16px;
+    height: 264.38px;
+}
+
+.wrapper {
+    position: relative;
+    width: 470px;
+    height: 258px;
+    overflow: hidden;
+}
+
+.img {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 470px;
+    max-height: 264.38px;
+    border-radius: 4px;
+}
